+++ src/main/java/de/kosit/validationtool/impl/tasks/ComputeAcceptanceAction.java
@@ -2,15 +2,19 @@ package de.kosit.validationtool.impl.tasks;
 
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
+import org.oclc.purl.dsdl.svrl.FailedAssert;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import de.kosit.validationtool.api.AcceptRecommendation;
 
+import net.sf.saxon.s9api.SaxonApiException;
 import net.sf.saxon.s9api.XPathSelector;
 
 /**
- * Berechnet die Akzeptanz-Empfehlung gemäß konfigurierten 'acceptMatch' des aktuellen Szenarios.
+ * Computes a {@link AcceptRecommendation} for this instance. This is either based on an 'acceptMatch'-configuration of
+ * the active scenario or based on overall evaluation about schema and semantic (schematron) correctness of the
  * 
  * @author Andreas Penski
  */
@@ -20,23 +24,49 @@ public class ComputeAcceptanceAction implements CheckAction {
 
     @Override
     public void check(final Bag results) {
-        final String acceptMatch = results.getScenarioSelectionResult().getObject().getAcceptMatch();
-        if (isNotBlank(acceptMatch)) {
-
-            try {
-
-                final XPathSelector selector = results.getScenarioSelectionResult().getObject().getAcceptSelector();
-                selector.setContextItem(results.getReport());
-                results.setAcceptStatus(selector.effectiveBooleanValue() ? AcceptRecommendation.ACCEPTABLE : AcceptRecommendation.REJECT);
-            } catch (final Exception e) {
-                log.error("Fehler bei Evaluierung des Accept-Status: {}", e.getMessage(), e);
+        if (preCondtionsMatch(results)) {
+            final String acceptMatch = results.getScenarioSelectionResult().getObject().getAcceptMatch();
+            if (results.getSchemaValidationResult().isValid() && isNotBlank(acceptMatch)) {
+                evaluateAcceptanceMatch(results);
+            } else {
+                evaluateSchemaAndSchematron(results);
             }
+        } else {
+            results.setAcceptStatus(AcceptRecommendation.REJECT);
         }
     }
 
-    @Override
-    public boolean isSkipped(final Bag results) {
-        return results.getReport() == null;
+    private void evaluateSchemaAndSchematron(final Bag results) {
+        if (results.getSchemaValidationResult().isValid() && isSchematronValid(results)) {
+            results.setAcceptStatus(AcceptRecommendation.ACCEPTABLE);
+        } else {
+            results.setAcceptStatus(AcceptRecommendation.REJECT);
+        }
+    }
+
+    private boolean isSchematronValid(final Bag results) {
+        return !hasSchematronErrors(results);
+    }
+
+    private boolean hasSchematronErrors(final Bag results) {
+        return results.getReportInput().getValidationResultsSchematron().stream().map(e -> e.getResults().getSchematronOutput())
+                .flatMap(e -> e.getActivePatternAndFiredRuleAndFailedAssert().stream()).anyMatch(FailedAssert.class::isInstance);
+    }
+
+    private static void evaluateAcceptanceMatch(final Bag results) {
+        try {
+            final XPathSelector selector = results.getScenarioSelectionResult().getObject().getAcceptSelector();
+            selector.setContextItem(results.getReport());
+            results.setAcceptStatus(selector.effectiveBooleanValue() ? AcceptRecommendation.ACCEPTABLE : AcceptRecommendation.REJECT);
+        } catch (final SaxonApiException e) {
+            final String msg = "Error evaluating accept recommendation: %s";
+            log.error(msg);
+            results.addProcessingError(msg);
+        }
+    }
+
+    private static boolean preCondtionsMatch(final Bag results) {
+        return results.getReport() != null && results.getSchemaValidationResult() != null && results.getScenarioSelectionResult() != null;
     }
 
 }