+++ src/test/java/de/kosit/validationtool/impl/DefaultCheckTest.java
@@ -74,8 +74,9 @@ public class DefaultCheckTest {
         final Result doc = this.implementation.checkInput(read(Simple.FOO));
         assertThat(doc).isNotNull();
         assertThat(doc.getReport()).isNotNull();
+        // happy case has schematron errors !??
         assertThat(doc.isAcceptable()).isFalse();
-        assertThat(doc.getAcceptRecommendation()).isEqualTo(AcceptRecommendation.UNDEFINED);
+        assertThat(doc.getAcceptRecommendation()).isEqualTo(AcceptRecommendation.REJECT);
     }
 
     @Test